refactor: provide `ServeHealthCheck` in genericapiserver package
