+++ internal/pump/run.go
@@ -5,13 +5,13 @@
 package pump
 
 import (
+	genericapiserver "github.com/marmotedu/iam/internal/pkg/server"
 	"github.com/marmotedu/iam/internal/pump/config"
-	"github.com/marmotedu/iam/internal/pump/server"
 )
 
 // Run runs the specified pump server. This should never exit.
 func Run(cfg *config.Config, stopCh <-chan struct{}) error {
-	go server.ServeHealthCheck(cfg.HealthCheckPath, cfg.HealthCheckAddress)
+	go genericapiserver.ServeHealthCheck(cfg.HealthCheckPath, cfg.HealthCheckAddress)
 
 	server, err := createPumpServer(cfg)
 	if err != nil {
